SpdxExpression: Remove the default ConsoleErrorListener
Ensure that always only our custom error listener is set to avoid errors
being written to the console. This e.g. avoids

    line 1:4 token recognition error at: ':'

being written to stderr when parsing the URL

    http://jquery.org/license

as an SpdxExpression.
